+++ internal/service/tp_object_service.go
@@ -6,6 +6,8 @@ import (
 	"backend-v2/model"
 	"fmt"
 	"os"
+	"path/filepath"
+	"time"
 
 	"github.com/xuri/excelize/v2"
 )
@@ -15,7 +17,8 @@ type tpObjectService struct {
 	workerRepo            repository.IWorkerRepository
 	objectSupervisorsRepo repository.IObjectSupervisorsRepository
 	objectTeamsRepo       repository.IObjectTeamsRepository
-  objectRepo repository.IObjectRepository
+	objectRepo            repository.IObjectRepository
+	teamRepo              repository.ITeamRepository
 }
 
 func InitTPObjectService(
@@ -23,35 +26,39 @@ func InitTPObjectService(
 	workerRepo repository.IWorkerRepository,
 	objectSupervisorsRepo repository.IObjectSupervisorsRepository,
 	objectTeamsRepo repository.IObjectTeamsRepository,
-  objectRepo repository.IObjectRepository,
+	objectRepo repository.IObjectRepository,
+	teamRepo repository.ITeamRepository,
 ) ITPObjectService {
 	return &tpObjectService{
 		tpObjectRepo:          tpObjectRepo,
 		workerRepo:            workerRepo,
 		objectSupervisorsRepo: objectSupervisorsRepo,
 		objectTeamsRepo:       objectTeamsRepo,
-    objectRepo: objectRepo,
+		objectRepo:            objectRepo,
+		teamRepo:              teamRepo,
 	}
 }
 
 type ITPObjectService interface {
-  GetAllOnlyObjects(projectID uint) ([]model.Object, error)
-	GetPaginated(page, limit int, projectID uint) ([]dto.TPObjectPaginated, error)
-	Count(projectID uint) (int64, error)
+	GetAllOnlyObjects(projectID uint) ([]model.Object, error)
+	GetPaginated(page, limit int, filter dto.TPObjectSearchParameters) ([]dto.TPObjectPaginated, error)
+	Count(filter dto.TPObjectSearchParameters) (int64, error)
 	Create(data dto.TPObjectCreate) (model.TP_Object, error)
 	Update(data dto.TPObjectCreate) (model.TP_Object, error)
 	Delete(id, projectID uint) error
-	TemplateFile(filepath string, projectID uint) error
+	TemplateFile(filePath string, projectID uint) (string, error)
 	Import(projectID uint, filepath string) error
+	GetObjectNamesForSearch(projectID uint) ([]dto.DataForSelect[string], error)
+	Export(projectID uint) (string, error)
 }
 
-func(service *tpObjectService) GetAllOnlyObjects(projectID  uint) ([]model.Object, error) {
-  return service.objectRepo.GetAllObjectBasedOnType(projectID, "tp_objects") 
+func (service *tpObjectService) GetAllOnlyObjects(projectID uint) ([]model.Object, error) {
+	return service.objectRepo.GetAllObjectBasedOnType(projectID, "tp_objects")
 }
 
-func (service *tpObjectService) GetPaginated(page, limit int, projectID uint) ([]dto.TPObjectPaginated, error) {
+func (service *tpObjectService) GetPaginated(page, limit int, filter dto.TPObjectSearchParameters) ([]dto.TPObjectPaginated, error) {
 
-	data, err := service.tpObjectRepo.GetPaginated(page, limit, projectID)
+	data, err := service.tpObjectRepo.GetPaginated(page, limit, filter)
 	if err != nil {
 		return []dto.TPObjectPaginated{}, err
 	}
@@ -75,7 +82,6 @@ func (service *tpObjectService) GetPaginated(page, limit int, projectID uint) ([
 			Status:           object.Status,
 			Model:            object.Model,
 			VoltageClass:     object.VoltageClass,
-			Nourashes:        object.Nourashes,
 			Supervisors:      supervisorNames,
 			Teams:            teamNumbers,
 		})
@@ -84,8 +90,8 @@ func (service *tpObjectService) GetPaginated(page, limit int, projectID uint) ([
 	return result, nil
 }
 
-func (service *tpObjectService) Count(projectID uint) (int64, error) {
-	return service.tpObjectRepo.Count(projectID)
+func (service *tpObjectService) Count(filter dto.TPObjectSearchParameters) (int64, error) {
+	return service.tpObjectRepo.Count(filter)
 }
 
 func (service *tpObjectService) Create(data dto.TPObjectCreate) (model.TP_Object, error) {
@@ -100,32 +106,51 @@ func (service *tpObjectService) Delete(id, projectID uint) error {
 	return service.tpObjectRepo.Delete(id, projectID)
 }
 
-func (service *tpObjectService) TemplateFile(filepath string, projectID uint) error {
+func (service *tpObjectService) TemplateFile(filePath string, projectID uint) (string, error) {
 
-	f, err := excelize.OpenFile(filepath)
+	f, err := excelize.OpenFile(filePath)
 	if err != nil {
 		f.Close()
-		return fmt.Errorf("Не смог открыть шаблонный файл: %v", err)
+		return "", fmt.Errorf("Не смог открыть шаблонный файл: %v", err)
 	}
 
 	sheetName := "Супервайзеры"
 	allSupervisors, err := service.workerRepo.GetByJobTitleInProject("Супервайзер", projectID)
 	if err != nil {
 		f.Close()
-		return fmt.Errorf("Данные супервайзеров недоступны: %v", err)
+		return "", fmt.Errorf("Данные супервайзеров недоступны: %v", err)
 	}
 
 	for index, supervisor := range allSupervisors {
 		f.SetCellValue(sheetName, "A"+fmt.Sprint(index+2), supervisor.Name)
 	}
 
-	if err := f.Save(); err != nil {
-		return fmt.Errorf("Не удалось обновить шаблон с новыми данными: %v", err)
+	allTeams, err := service.teamRepo.GetAll(projectID)
+	if err != nil {
+		f.Close()
+		return "", fmt.Errorf("Данны бригад не доступны: %v", err)
 	}
 
-	f.Close()
+	teamSheetName := "Бригады"
+	for index, team := range allTeams {
+		f.SetCellStr(teamSheetName, "A"+fmt.Sprint(index+2), team.Number)
+	}
 
-	return nil
+	currentTime := time.Now()
+	temporaryFileName := fmt.Sprintf(
+		"Шаблон импорта ТП - %s.xlsx",
+		currentTime.Format("02-01-2006"),
+	)
+	temporaryFilePath := filepath.Join("./pkg/excels/temp/", temporaryFileName)
+	if err := f.SaveAs(temporaryFilePath); err != nil {
+		return "", fmt.Errorf("Не удалось обновить шаблон с новыми данными: %v", err)
+	}
+
+	if err := f.Close(); err != nil {
+		return "", err
+	}
+
+	return temporaryFilePath, nil
 }
 
 func (service *tpObjectService) Import(projectID uint, filepath string) error {
@@ -137,7 +162,7 @@ func (service *tpObjectService) Import(projectID uint, filepath string) error {
 		return fmt.Errorf("Не смог открыть файл: %v", err)
 	}
 
-	sheetName := "Импорт"
+	sheetName := "ТП"
 	rows, err := f.GetRows(sheetName)
 	if err != nil {
 		f.Close()
@@ -151,9 +176,7 @@ func (service *tpObjectService) Import(projectID uint, filepath string) error {
 		return fmt.Errorf("Файл не имеет данных")
 	}
 
-	objects := []model.Object{}
-	tps := []model.TP_Object{}
-	supervisorIDs := []uint{}
+	tps := []dto.TPObjectImportData{}
 	index := 1
 	for len(rows) > index {
 		object := model.Object{
@@ -177,44 +200,62 @@ func (service *tpObjectService) Import(projectID uint, filepath string) error {
 			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке B%d: %v", index+1, err)
 		}
 
-		supervisorName, err := f.GetCellValue(sheetName, "C"+fmt.Sprint(index+1))
-		if err != nil {
-			f.Close()
-			os.Remove(filepath)
-			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке C%d: %v", index+1, err)
-		}
-
-		supervisorWorker, err := service.workerRepo.GetByName(supervisorName)
-		if err != nil {
-			f.Close()
-			os.Remove(filepath)
-			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке C%d: %v", index+1, err)
-		}
-
-		tp.Model, err = f.GetCellValue(sheetName, "D"+fmt.Sprint(index+1))
+		tp.Model, err = f.GetCellValue(sheetName, "C"+fmt.Sprint(index+1))
 		if err != nil {
 			f.Close()
 			os.Remove(filepath)
 			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке D%d: %v", index+1, err)
 		}
 
-		tp.VoltageClass, err = f.GetCellValue(sheetName, "E"+fmt.Sprint(index+1))
+		tp.VoltageClass, err = f.GetCellValue(sheetName, "D"+fmt.Sprint(index+1))
 		if err != nil {
 			f.Close()
 			os.Remove(filepath)
 			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке E%d: %v", index+1, err)
 		}
 
-		tp.Nourashes, err = f.GetCellValue(sheetName, "F"+fmt.Sprint(index+1))
+		supervisorName, err := f.GetCellValue(sheetName, "E"+fmt.Sprint(index+1))
 		if err != nil {
 			f.Close()
 			os.Remove(filepath)
-			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке F%d: %v", index+1, err)
+			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке G%d: %v", index+1, err)
+		}
+		supervisorWorker := model.Worker{}
+		if supervisorName != "" {
+			supervisorWorker, err = service.workerRepo.GetByName(supervisorName)
+			if err != nil {
+				f.Close()
+				os.Remove(filepath)
+				return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке G%d: %v", index+1, err)
+			}
 		}
 
-		supervisorIDs = append(supervisorIDs, supervisorWorker.ID)
-		objects = append(objects, object)
-		tps = append(tps, tp)
+		teamNumber, err := f.GetCellValue(sheetName, "F"+fmt.Sprint(index+1))
+		if err != nil {
+			f.Close()
+			os.Remove(filepath)
+			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке H%d: %v", index+1, err)
+		}
+		team := model.Team{}
+		if teamNumber != "" {
+			team, err = service.teamRepo.GetByNumber(teamNumber)
+			if err != nil {
+				f.Close()
+				os.Remove(filepath)
+				return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке H%d: %v", index+1, err)
+			}
+		}
+
+		tps = append(tps, dto.TPObjectImportData{
+			Object: object,
+			TP:     tp,
+			ObjectSupervisors: model.ObjectSupervisors{
+				SupervisorWorkerID: supervisorWorker.ID,
+			},
+			ObjectTeam: model.ObjectTeams{
+				TeamID: team.ID,
+			},
+		})
 		index++
 	}
 
@@ -226,10 +267,93 @@ func (service *tpObjectService) Import(projectID uint, filepath string) error {
 		return fmt.Errorf("Ошибка при удалении временного файла: %v", err)
 	}
 
-	_, err = service.tpObjectRepo.CreateInBatches(objects, tps, supervisorIDs)
-	if err != nil {
+	if err := service.tpObjectRepo.CreateInBatches(tps); err != nil {
 		return err
 	}
 
 	return nil
+}
+
+func (service *tpObjectService) GetObjectNamesForSearch(projectID uint) ([]dto.DataForSelect[string], error) {
+	return service.tpObjectRepo.GetObjectNamesForSearch(projectID)
+}
+
+func (service *tpObjectService) Export(projectID uint) (string, error) {
+	tpTempalteFilePath := filepath.Join("./pkg/excels/templates/", "Шаблон для импорта ТП.xlsx")
+	f, err := excelize.OpenFile(tpTempalteFilePath)
+	if err != nil {
+		f.Close()
+		return "", fmt.Errorf("Не смог открыть файл: %v", err)
+	}
+	sheetName := "ТП"
+	startingRow := 2
+
+	tpCount, err := service.tpObjectRepo.Count(dto.TPObjectSearchParameters{ProjectID: projectID})
+	if err != nil {
+		return "", err
+	}
+	limit := 100
+	page := 1
+
+	for tpCount > 0 {
+		tps, err := service.tpObjectRepo.GetPaginated(page, limit, dto.TPObjectSearchParameters{ProjectID: projectID})
+		if err != nil {
+			return "", err
+		}
+
+		for index, tp := range tps {
+			supervisorNames, err := service.objectSupervisorsRepo.GetSupervisorsNameByObjectID(tp.ObjectID)
+			if err != nil {
+				return "", err
+			}
+
+			teamNumbers, err := service.objectTeamsRepo.GetTeamsNumberByObjectID(tp.ObjectID)
+			if err != nil {
+				return "", err
+			}
+
+			f.SetCellStr(sheetName, "A"+fmt.Sprint(startingRow+index), tp.Name)
+			f.SetCellStr(sheetName, "B"+fmt.Sprint(startingRow+index), tp.Status)
+			f.SetCellStr(sheetName, "C"+fmt.Sprint(startingRow+index), tp.Model)
+			f.SetCellStr(sheetName, "D"+fmt.Sprint(startingRow+index), tp.VoltageClass)
+
+			supervisorsCombined := ""
+			for index, supervisor := range supervisorNames {
+				if index == 0 {
+					supervisorsCombined += supervisor
+					continue
+				}
+
+				supervisorsCombined += ", " + supervisor
+			}
+			f.SetCellStr(sheetName, "E"+fmt.Sprint(startingRow+index), supervisorsCombined)
+
+			teamNumbersCombined := ""
+			for index, teamNumber := range teamNumbers {
+				if index == 0 {
+					teamNumbersCombined += teamNumber
+					continue
+				}
+
+				teamNumbersCombined += ", " + teamNumber
+			}
+			f.SetCellStr(sheetName, "F"+fmt.Sprint(startingRow+index), teamNumbersCombined)
+		}
+
+		startingRow = page*limit + 2
+		page++
+		tpCount -= int64(limit)
+	}
+
+	currentTime := time.Now()
+	exportFileName := fmt.Sprintf(
+		"Экспорт ТП - %s.xlsx",
+		currentTime.Format("02-01-2006"),
+	)
+	exportFilePath := filepath.Join("./pkg/excels/temp/", exportFileName)
+	if err := f.SaveAs(exportFilePath); err != nil {
+		return "", err
+	}
+
+	return exportFileName, nil
 }
